fix: 최신 계약 재확인 반영 — 비율 표기, 에러 data:null, 챗봇 context/role/tool 파라미터

+++ app/schemas/common.py
@@ -20,7 +20,7 @@ class Evidence(Contract):
 
 class SalesSummary(Contract):
     netSales: int
-    vsPrevPeriod: int
+    vsPrevPeriod: float
 
 
 class HourlyPoint(Contract):
@@ -31,8 +31,8 @@ class HourlyPoint(Contract):
 
 class CategoryPoint(Contract):
     name: str
-    share: int
-    vsPrevPeriod: int
+    share: float
+    vsPrevPeriod: float
 
 
 class Metrics(Contract):
